+++ ebird-api/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Revision history for ebird-api
 
+## 0.2.0.0 -- 2023-12-16
+
+* Rename subregion items so subregion is one word.
+
 ## 0.1.0.0 -- 2023-07-30
 
 * First version. Released on an unsuspecting world.