+++ ebird-client/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Revision history for ebird-client
 
+## 0.2.0.0 -- 2023-12-16
+
+* Rename subregion items so that subregion is one word.
+* Bump `ebird-api` constraint.
+
 ## 0.1.0.0 -- 2023-07-30
 
 * First version. Released on an unsuspecting world.